Use SparseArrays in UniversalKh ===============================
While the hard computations for UniversalKh is done by the underlying Java
program, the Mathematica wrapper (which just does some elementary
manipulations of the resulting chain complex) actually takes the vast bulk
of the time for larger examples.  This switch to Mathematica's
SparseArrays gives a huge speedup (100 times or more in some cases).